Refactor cyral_policy_rule and fix error in cyral_service_account
