+++ cyral/internal/policy/rule/constants.go
@@ -1,0 +1,5 @@
+package rule
+
+const (
+	resourceName = "cyral_policy_rule"
+)
