+++ cyral/internal/sidecar/listener/resource.go
@@ -24,7 +24,7 @@ var resourceContextHandler = core.DefaultContextHandler{
 			c.ControlPlane,
 			d.Get(utils.SidecarIDKey).(string))
 	},
-	GetPutDeleteURLFactory: func(d *schema.ResourceData, c *client.Client) string {
+	ReadUpdateDeleteURLFactory: func(d *schema.ResourceData, c *client.Client) string {
 		return fmt.Sprintf("https://%s/v1/sidecars/%s/listeners/%s",
 			c.ControlPlane,
 			d.Get(utils.SidecarIDKey).(string),
