+++ cyral/provider/schema_loader.go
@@ -13,6 +13,7 @@ import (
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/integration/slack"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/integration/teams"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/policy"
+	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/policy/rule"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository/accessgateway"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository/accessrules"
@@ -51,6 +52,7 @@ func packagesSchemas() []core.PackageSchema {
 		pagerduty.PackageSchema(),
 		policy.PackageSchema(),
 		repository.PackageSchema(),
+		rule.PackageSchema(),
 		samlcertificate.PackageSchema(),
 		serviceaccount.PackageSchema(),
 		sidecar.PackageSchema(),
